deepseek-api/requirements.txt
torch==2.1.0
transformers==4.35.2
accelerate==0.24.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
uuid==1.30
numpy==1.24.3
pandas==2.1.3
sentencepiece==0.1.99
protobuf==4.25.1
requests==2.31.0
scipy==1.11.4